- Removed a CPU intensive AI calculation function which was used by HAM (thx to Headrock) - Fixed some compilation warnings
git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk/GameSource/ja2_v1.13/Build@2655 3b4a5df2-a311-0410-b5c6-a8a6f20db521

+++ Tactical/Overhead.cpp
@@ -7267,7 +7267,7 @@ void HandleSuppressionFire( UINT8 ubTargetedMerc, UINT8 ubCausedAttacker )
 			{
 				if (ubPointsLost > APBPConstants[AP_MAX_SUPPRESSED])
 				{
-					ubPointsLost = APBPConstants[AP_MAX_SUPPRESSED];
+					ubPointsLost = (UINT8)APBPConstants[AP_MAX_SUPPRESSED];
 				}
 			}
